Adds asClass parameter for Component annotation
This parameter permits to force the class used to register the component,
for example a Component can register itself on one of its implemented interface:
````
@Component(asClass = MyInterface.class)
public class MyImplementation implements MyInterface {
...
````

It was doable using Provides annotation in a module, but was missing when the Component annotation
was used.

WIP: Manage conditional component, they use another template, and method in the annotation processor.